turn initial graph building step into first pipeline step
