+++ src/requirements/requirement-frontend-computation.ts
@@ -7,7 +7,7 @@ import {
   computeFulfillmentCoursesAndStatistics,
 } from './requirement-frontend-utils';
 import RequirementFulfillmentGraph from './graph';
-import buildRequirementFulfillmentGraphFromUserData from './graph/requirement-graph-builder-from-user-data';
+import buildRequirementFulfillmentGraphFromUserData from './requirement-graph-builder-from-user-data';
 import featureFlagCheckers from '../feature-flags';
 
 /**
